requirements.txt
Django==5.2.4
djangorestframework==3.14.0
gunicorn==21.2.0
whitenoise==6.5.0
python-decouple==3.8
psycopg2-binary==2.9.7
dj-database-url==2.1.0